NOTE: compatibility update for root version <5.30; requires rerunning of all mscw_energy files, no backwards compatibility! (tables are ok); changed NTtype from UInt_t to Int_t; fixed finder for constant variables in TMVA trainer; added additionally checks for consistent files for effective area calculation
git-svn-id: https://svnsrv.desy.de/desy/EVNDISP/trunk@1031 766874ec-0625-4018-90f7-740f284a29e6

+++ src/VGammaHadronCuts.cpp
@@ -2084,7 +2084,7 @@ bool VNTelTypeCut::test( CData *c )
    unsigned int ntel_type = 0;
    for( unsigned int i = 0; i < fTelType_counter.size(); i++ )
    {
-      if( fTelType_counter[i] < c->NTtype ) ntel_type += c->NImages_Ttype[fTelType_counter[i]];
+      if( (int)fTelType_counter[i] < c->NTtype ) ntel_type += c->NImages_Ttype[fTelType_counter[i]];
    }
 // OBS! >=
    if( ntel_type >= fNTelType_min ) return true;
